修复了打包大小不正确的BUG；修复了目标目录不存在时报错的BUG；修复了进度显示的BUG

+++ ArchiveMaster.Module.WriteOnceArchive/Configs/PackingConfig.cs
@@ -19,7 +19,7 @@ public partial class PackingConfig : ConfigBase
     private PackingType packingType = PackingType.Copy;
 
     [ObservableProperty]
-    private double packageSizeMB = 23500;
+    private long packageSizeMB = 23500;
 
     [ObservableProperty]
     private string previousPackageInfoFiles;
